Rework tests and some additional Config class work
