--- tests/test_validator.py
@@ -1,31 +1,0 @@
-import pytest
-
-from lambdas import validator
-from lambdas.config import configure_sentry
-
-
-def test_validator_configures_sentry_if_dsn_present(caplog, monkeypatch):
-    monkeypatch.setenv("SENTRY_DSN", "https://1234567890@00000.ingest.sentry.io/123456")
-    configure_sentry()
-    assert (
-        "Sentry DSN found, exceptions will be sent to Sentry with env=test" in caplog.text
-    )
-
-
-def test_validator_doesnt_configure_sentry_if_dsn_not_present(caplog, monkeypatch):
-    monkeypatch.delenv("SENTRY_DSN", raising=False)
-    configure_sentry()
-    assert "No Sentry DSN found, exceptions will not be sent to Sentry" in caplog.text
-
-
-def test_lambda_handler_missing_workspace_env_raises_error(monkeypatch):
-    monkeypatch.delenv("WORKSPACE", raising=False)
-    with pytest.raises(
-        OSError,
-        match="Missing required environment variables: WORKSPACE",
-    ):
-        validator.lambda_handler({})
-
-
-def test_validator():
-    assert validator.lambda_handler({}) == "You have successfully called this lambda!"
